Add files via upload
This is my very first application in Qt Quick Qml language. It has some mistakes and overflows but in general this is a front end view of my application. I did not code back end of this application but tried to add some features on it. At that time i wasn't aware about Object Oriented clearly, consequently my code might be little complicated for you.Further more, i took 2 week lessons for Qt Quick Qml for learning, i did this application in 20 days. My first learning about this language and first attempt for this application.